hdl: nios: devices_rfic: split into _cmds
Move functions around to clean up things and make files shorter.

Pass state pointer around to reduce global variable scope.

Clean up some magic numbers and unnecessarily-verbose strings
and function declarations.
