+++ hdl/fpga/platforms/common/bladerf/software/bladeRF_nios/src/devices.h
@@ -581,6 +581,7 @@ uint8_t rx_trigger_ctl_read(void);
  */
 void agc_dc_corr_write(uint16_t addr, uint16_t value);
 
+#ifdef BLADERF_NIOS_LIBAD936X
 /**
  * Initialize the RFIC command handler. Run once at startup.
  */
@@ -614,7 +615,7 @@ bool rfic_command_write(uint16_t addr, uint64_t data);
  * @return bool (true = success)
  */
 bool rfic_command_read(uint16_t addr, uint64_t *data);
-
+#endif  // BLADERF_NIOS_LIBAD936X
 
 /* A number of rountines define here are implemented as just a register
  * access, where incurring function call overhead is wasteful. Therefore,
